refactor: move manager to other namespace
